# packages/contracts/openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Sophia application API",
    "version": "0.4.0-design",
    "description": "Sophia-owned binding contract. Not the Omnigent OpenAPI and not a claim that these routes are deployed."
  },
  "paths": {
...
    "/v1/media/assignments": {
      "get": {
        "operationId": "mediaAssignments",
        "summary": "The exchanges the bridge must serve, with a room token for each; waits up to waitMs for a change after `after`.",
        "tags": [
          "media"
        ],
        "security": [
          {
            "mediaBridgeCapability": []
          }
        ],
        "parameters": [
          {
            "name": "after",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "pattern": "^[0-9a-f]{64}$"
            }
          },
          {
            "name": "waitMs",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^(0|[1-9][0-9]{0,4})$"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MediaAssignmentBatch"
                }
              }
            }
          },
          "401": {
            "description": "No or unknown media-bridge capability"
          },
          "409": {
            "description": "Stale: the exchange or epoch moved on",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "Request fails the declared schema",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Dependent service unavailable"
          }
        },
        "x-sophia-implementation": "contract_only",
        "x-sophia-release-goal": "S1-05A"
      }
    },
    "/v1/media/presence": {
      "post": {
        "operationId": "mediaPresence",
        "summary": "The bridge's view of one room: provider state and the trusted participant list. A guest in the room pauses a project-aware exchange.",
        "tags": [
          "media"
        ],
        "security": [
          {
            "mediaBridgeCapability": []
          }
        ],
        "parameters": [],
        "responses": {
          "204": {
            "description": "Recorded"
          },
          "401": {
            "description": "No or unknown media-bridge capability"
          },
          "409": {
            "description": "Stale: the exchange or epoch moved on",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "Request fails the declared schema",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Dependent service unavailable"
          }
        },
        "x-sophia-implementation": "contract_only",
        "x-sophia-release-goal": "S1-05A",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MediaPresenceReport"
              }
            }
          }
        }
      }
    },
    "/v1/media/quiesce-acks": {
      "post": {
        "operationId": "mediaQuiesceAck",
        "summary": "The bridge confirms it closed Google input and cleared Sophia's output before a guest is admitted.",
        "tags": [
          "media"
        ],
        "security": [
          {
            "mediaBridgeCapability": []
          }
        ],
        "parameters": [],
        "responses": {
          "204": {
            "description": "Recorded"
          },
          "401": {
            "description": "No or unknown media-bridge capability"
          },
          "409": {
            "description": "Stale: the exchange or epoch moved on",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "Request fails the declared schema",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Dependent service unavailable"
          }
        },
        "x-sophia-implementation": "contract_only",
        "x-sophia-release-goal": "S1-05A",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MediaQuiesceAck"
              }
            }
          }
        }
      }
    },
    "/v1/media/holder": {
      "post": {
        "operationId": "mediaHolderEvent",
        "summary": "The input holder left the room (input pauses at once) or stayed gone past the reconnect grace (the floor is cleared by compare-and-set against that holder and epoch).",
        "tags": [
          "media"
        ],
        "security": [
          {
            "mediaBridgeCapability": []
          }
        ],
        "parameters": [],
        "responses": {
          "204": {
            "description": "Recorded"
          },
          "401": {
            "description": "No or unknown media-bridge capability"
          },
          "409": {
            "description": "Stale: the exchange or epoch moved on",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "Request fails the declared schema",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Dependent service unavailable"
          }
        },
        "x-sophia-implementation": "contract_only",
        "x-sophia-release-goal": "S1-05A",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MediaHolderEvent"
              }
            }
          }
        }
      }
    },
    "/v1/media/announced": {
      "post": {
        "operationId": "mediaAnnounced",
        "summary": "The bridge told the room about a finished background result, once: it is recorded so no reconnect or restart announces it again.",
        "tags": [
          "media"
        ],
        "security": [
          {
            "mediaBridgeCapability": []
          }
        ],
        "parameters": [],
        "responses": {
          "204": {
            "description": "Recorded"
          },
          "401": {
            "description": "No or unknown media-bridge capability"
          },
          "409": {
            "description": "Stale: the exchange or epoch moved on",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "Request fails the declared schema",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Dependent service unavailable"
          }
        },
        "x-sophia-implementation": "contract_only",
        "x-sophia-release-goal": "S1-05A",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MediaAnnounced"
              }
            }
          }
        }
      }
    },
    "/v1/media/tool-calls": {
      "post": {
        "operationId": "mediaToolCall",
        "summary": "One Gemini Live tool call, executed for the speaker bound to its input epoch and under that speaker's role. Idempotent per exchange, connection generation and call id.",
        "tags": [
          "media"
        ],
        "security": [
          {
            "mediaBridgeCapability": []
          }
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MediaToolResult"
                }
              }
            }
          },
          "401": {
            "description": "No or unknown media-bridge capability"
          },
          "409": {
            "description": "Stale: the exchange or epoch moved on",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "Request fails the declared schema",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Dependent service unavailable"
          }
        },
        "x-sophia-implementation": "contract_only",
        "x-sophia-release-goal": "S1-05A",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MediaToolCall"
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "supabaseBearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "runtimeCapability": {
        "type": "http",
        "scheme": "bearer",
        "description": "A runtime capability: an opaque secret held by the execution host and stored by the service only as SHA-256. It is bound to one project, executor resource and runtime unit, and it is not a member identity."
      },
      "mediaBridgeCapability": {
        "type": "http",
        "scheme": "bearer",
        "description": "The media bridge's capability: an opaque secret on the bridge host; the API holds only its SHA-256. It names no member and grants no member authority: tool calls act for the speaker the exchange's input epoch binds, under that speaker's own role."
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "requestId": {
            "type": "string",
            "format": "uuid"
          },
          "retry": {
            "type": "string",
            "enum": [
              "safe_read",
              "same_admission_key",
              "reconcile_first",
              "reauthorize",
              "never"
            ]
          }
        },
        "required": [
          "code",
          "message",
          "requestId",
          "retry"
        ]
      },
...
      "MediaRoomToken": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "serverUrl": {
            "type": "string"
          },
          "token": {
            "type": "string"
          },
          "expiresAt": {
            "type": "string"
          }
        },
        "required": [
          "serverUrl",
          "token",
          "expiresAt"
        ]
      },
      "MediaAssignment": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "exchangeId": {
            "type": "string",
            "pattern": "^[0-9a-f-]{36}$"
          },
          "projectId": {
            "type": "string",
            "pattern": "^[0-9a-f-]{36}$"
          },
          "roomId": {
            "type": "string",
            "pattern": "^[0-9a-f-]{36}$"
          },
          "state": {
            "type": "string",
            "enum": [
              "open",
              "paused"
            ]
          },
          "pauseReason": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "guest",
                  "holder_left"
                ]
              },
              {
                "type": "null"
              }
            ]
          },
          "inputEpoch": {
            "type": "integer",
            "minimum": 1,
            "maximum": 9007199254740991
          },
          "inputActorId": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^[0-9a-f-]{36}$"
              },
              {
                "type": "null"
              }
            ]
          },
          "playbackEpoch": {
            "type": "integer",
            "minimum": 1,
            "maximum": 9007199254740991
          },
          "observationEpoch": {
            "type": "integer",
            "minimum": 1,
            "maximum": 9007199254740991
          },
          "allowVision": {
            "type": "boolean"
          },
          "looking": {
            "anyOf": [
              {
                "type": "object",
                "additionalProperties": false,
                "properties": {
                  "participantIdentity": {
                    "type": "string",
                    "pattern": "^[A-Za-z0-9][A-Za-z0-9_.:-]{0,127}$"
                  },
                  "source": {
                    "type": "string",
                    "enum": [
                      "screen",
                      "camera"
                    ]
                  }
                },
                "required": [
                  "participantIdentity",
                  "source"
                ]
              },
              {
                "type": "null"
              }
            ]
          },
          "roomRevision": {
            "type": "integer",
            "minimum": 1,
            "maximum": 9007199254740991
          },
          "quiesceRequestId": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^[0-9a-f-]{36}$"
              },
              {
                "type": "null"
              }
            ]
          },
          "roomToken": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/MediaRoomToken"
              },
              {
                "type": "null"
              }
            ]
          },
          "results": {
            "type": "array",
            "maxItems": 20,
            "items": {
              "type": "object",
              "additionalProperties": false,
              "properties": {
                "taskId": {
                  "type": "string",
                  "pattern": "^[0-9a-f-]{36}$"
                },
                "resultRevision": {
                  "type": "integer",
                  "minimum": 1,
                  "maximum": 9007199254740991
                },
                "kind": {
                  "const": "draft_brief"
                }
              },
              "required": [
                "taskId",
                "resultRevision",
                "kind"
              ]
            }
          }
        },
        "required": [
          "exchangeId",
          "projectId",
          "roomId",
          "state",
          "pauseReason",
          "inputEpoch",
          "inputActorId",
          "playbackEpoch",
          "observationEpoch",
          "allowVision",
          "looking",
          "roomRevision",
          "quiesceRequestId",
          "roomToken",
          "results"
        ]
      },
      "MediaAssignmentBatch": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "assignments": {
            "type": "array",
            "maxItems": 200,
            "items": {
              "$ref": "#/components/schemas/MediaAssignment"
            }
          },
          "version": {
            "type": "string",
            "pattern": "^[0-9a-f]{64}$"
          }
        },
        "required": [
          "assignments",
          "version"
        ]
      },
      "MediaPresenceReport": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "roomId": {
            "type": "string",
            "pattern": "^[0-9a-f-]{36}$"
          },
          "exchangeId": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^[0-9a-f-]{36}$"
              },
              {
                "type": "null"
              }
            ]
          },
          "bridgeInstanceId": {
            "type": "string",
            "pattern": "^[A-Za-z0-9_-]{1,64}$"
          },
          "voice": {
            "type": "string",
            "enum": [
              "connecting",
              "ready",
              "recovering",
              "unavailable"
            ]
          },
          "reason": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 500
              },
              {
                "type": "null"
              }
            ]
          },
          "participants": {
            "type": "array",
            "maxItems": 200,
            "items": {
              "type": "object",
              "additionalProperties": false,
              "properties": {
                "identity": {
                  "type": "string",
                  "pattern": "^[A-Za-z0-9][A-Za-z0-9_.:-]{0,127}$"
                },
                "standing": {
                  "type": "string",
                  "enum": [
                    "admin",
                    "editor",
                    "viewer",
                    "guest",
                    "unknown"
                  ]
                }
              },
              "required": [
                "identity",
                "standing"
              ]
            }
          }
        },
        "required": [
          "roomId",
          "exchangeId",
          "bridgeInstanceId",
          "voice",
          "reason",
          "participants"
        ]
      },
      "MediaQuiesceAck": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "requestId": {
            "type": "string",
            "pattern": "^[0-9a-f-]{36}$"
          },
          "bridgeInstanceId": {
            "type": "string",
            "pattern": "^[A-Za-z0-9_-]{1,64}$"
          },
          "inputClosed": {
            "const": true
          },
          "outputCleared": {
            "const": true
          }
        },
        "required": [
          "requestId",
          "bridgeInstanceId",
          "inputClosed",
          "outputCleared"
        ]
      },
      "MediaHolderEvent": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "exchangeId": {
            "type": "string",
            "pattern": "^[0-9a-f-]{36}$"
          },
          "actorId": {
            "type": "string",
            "pattern": "^[0-9a-f-]{36}$"
          },
          "inputEpoch": {
            "type": "integer",
            "minimum": 1,
            "maximum": 9007199254740991
          },
          "event": {
            "type": "string",
            "enum": [
              "left",
              "gone"
            ]
          }
        },
        "required": [
          "exchangeId",
          "actorId",
          "inputEpoch",
          "event"
        ]
      },
      "MediaToolCall": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "exchangeId": {
            "type": "string",
            "pattern": "^[0-9a-f-]{36}$"
          },
          "connectionGeneration": {
            "type": "integer",
            "minimum": 1,
            "maximum": 9007199254740991
          },
          "callId": {
            "type": "string",
            "pattern": "^[A-Za-z0-9_.:-]{1,64}$"
          },
          "name": {
            "type": "string",
            "enum": [
              "project_status",
              "read_selected_source",
              "start_brief",
              "control_work"
            ]
          },
          "args": {
            "type": "object",
            "additionalProperties": true
          },
          "inputEpoch": {
            "type": "integer",
            "minimum": 1,
            "maximum": 9007199254740991
          },
          "actorId": {
            "type": "string",
            "pattern": "^[0-9a-f-]{36}$"
          }
        },
        "required": [
          "exchangeId",
          "connectionGeneration",
          "callId",
          "name",
          "args",
          "inputEpoch",
          "actorId"
        ]
      },
      "MediaAnnounced": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "exchangeId": {
            "type": "string",
            "pattern": "^[0-9a-f-]{36}$"
          },
          "taskId": {
            "type": "string",
            "pattern": "^[0-9a-f-]{36}$"
          },
          "resultRevision": {
            "type": "integer",
            "minimum": 1,
            "maximum": 9007199254740991
          }
        },
        "required": [
          "exchangeId",
          "taskId",
          "resultRevision"
        ]
      },
      "MediaToolResult": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "ok",
              "admitted",
              "refused",
              "clarify",
              "error"
            ]
          },
          "output": {
            "type": "object",
            "additionalProperties": true
          }
        },
        "required": [
          "status",
          "output"
        ]
      }
    }
  }
}